# cdc: run CheckLeader in a dedicate thread