<!--
Thank you for contributing to TiKV!

If you haven't already, please read TiKV's [CONTRIBUTING](https://github.com/tikv/tikv/blob/master/CONTRIBUTING.md) document.

If you're unsure about anything, just ask; somebody should be along to answer within a day or two.

PR Title Format:
1. module [, module2, module3]: what's changed
2. *: what's changed
-->

### What is changed and how it works?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and 
linking the relevant issues via the "close" or "ref".

For more info, check https://github.com/tikv/tikv/blob/master/CONTRIBUTING.md#linking-issues.

-->
Issue Number: Close #13774

What's Changed:

```commit-message
cdc: run CheckLeader in a dedicate thread
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

- Unit test
- Manual test (add detailed scripts or steps below)

TiCDC Latency is reduced from ~50s to ~10s, when there are lots of write workloads.

| Without This PR | With This PR |
| - | - |
| ![img_v2_ee229db2-9da3-4bf5-ba0a-0ff00a11baag](https://user-images.githubusercontent.com/2150711/201869965-4bd951be-6d46-4b5b-9a3a-54b2f8832dc7.jpg) | ![img_v2_467baa0f-0148-444d-86c6-6d02cfc799ag](https://user-images.githubusercontent.com/2150711/201870047-9437f168-ffe3-4db3-b554-b0f377f42919.jpg) |



### Release note <!-- bugfixes or new feature need a release note -->

```release-note
Run CheckLeader in a dedicate thread to stabilize TiCDC replication lag.
```